Another not integer. I know this won't ever happen but warnings

+++ src/mancala.h
@@ -24,7 +24,7 @@ class Board {
 	int their_mancala() const noexcept { return player == 0 ? 0 : 7; }
 	bool on_my_side(int n) const noexcept {
 		if (player == 0) return n > their_mancala() && n < my_mancala();
-		return n > their_mancala() && n < size();
+		return n > their_mancala() && n < static_cast<int>(size());
 	}
 
 public:
